# .specs/contracts/openapi.yaml
openapi: 3.0.4
info:
  title: Rampa Remittances API
  description: |
    API for Solana-based remittances app targeting Spain/Germany to LATAM corridor.
    
    ## Authentication Flow
    
    This API uses Web3Auth for user authentication with a token exchange pattern:
    
    1. **User Login**: User logs in via Web3Auth (Google, Apple, etc.) on the frontend
    2. **Token Exchange**: Frontend calls `/auth/web3auth/validate` with Web3Auth JWT token
    3. **API Token**: Backend validates Web3Auth token and returns our API JWT token
    4. **API Calls**: Frontend uses our API JWT token for all subsequent business API calls
    
    ## Security
    
    - All business endpoints require our API JWT token (Bearer authentication)
    - Web3Auth endpoints (`/auth/web3auth/*`) are public and don't require authentication
    - Health and inquiry endpoints are public
  version: 1.0.0
  contact:
    name: Rampa Team
    email: support@rampa.com
servers:
  - url: https://api.rampa.com
    description: Production server
  - url: https://staging-api.rampa.com
    description: Staging server
  - url: http://localhost:3001
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /visa-card:
    get:
      summary: Get VISA card details
      tags: [VISA Card]
      responses:
        '200':
          description: VISA card details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VISACardResponse'
        '404':
          description: No VISA card found
        '401':
          description: Unauthorized

    post:
      summary: Create VISA card
      tags: [VISA Card]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVISACardRequest'
      responses:
        '201':
          description: VISA card created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VISACardResponse'
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
        '403':
          description: Profile verification required for this operation
  /visa-card/all:
    get:
      summary: Get all VISA cards
      tags: [VISA Card]
      responses:
        '200':
          description: VISA cards retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VISACardResponse'
        '401':
          description: Unauthorized
  /visa-card/by-status/{status}:
    get:
      summary: Get VISA cards by status
      tags: [VISA Card]
      parameters:
        - name: status
          in: path
          required: true
          schema:
            type: string
            enum: [pending, active, suspended, cancelled]
      responses:
        '200':
          description: VISA cards retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VISACardResponse'
        '401':
          description: Unauthorized
  /visa-card/expired:
    get:
      summary: Get expired VISA cards
      tags: [VISA Card]
      responses:
        '200':
          description: Expired VISA cards retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VISACardResponse'
        '401':
          description: Unauthorized
  /visa-card/stats:
    get:
      summary: Get VISA card statistics
      tags: [VISA Card]
      responses:
        '200':
          description: VISA card statistics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalCards:
                    type: integer
                  activeCards:
                    type: integer
                  suspendedCards:
                    type: integer
                  cancelledCards:
                    type: integer
                  expiredCards:
                    type: integer
        '401':
          description: Unauthorized
  /visa-card/{id}:
    get:
      summary: Get specific VISA card
      tags: [VISA Card]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: VISA card retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VISACardResponse'
        '404':
          description: VISA card not found
        '401':
          description: Unauthorized

    put:
      summary: Update VISA card
      tags: [VISA Card]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVISACardRequest'
      responses:
        '200':
          description: VISA card updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VISACardResponse'
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
        '404':
          description: VISA card not found
  /visa-card/{id}/activate:
    post:
      summary: Activate VISA card
      tags: [VISA Card]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: VISA card activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
                  activatedAt:
                    type: string
                    format: date-time
        '401':
          description: Unauthorized
        '404':
          description: VISA card not found
  /visa-card/{id}/suspend:
    post:
      summary: Suspend VISA card
      tags: [VISA Card]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: VISA card suspended
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
        '401':
          description: Unauthorized
        '404':
          description: VISA card not found
  /visa-card/{id}/reactivate:
    post:
      summary: Reactivate VISA card
      tags: [VISA Card]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: VISA card reactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
        '401':
          description: Unauthorized
        '404':
          description: VISA card not found
  /visa-card/{id}/cancel:
    post:
      summary: Cancel VISA card
      tags: [VISA Card]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: VISA card cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
        '401':
          description: Unauthorized
        '404':
          description: VISA card not found
  /visa-card/{id}/update-balance:
    post:
      summary: Update VISA card balance
      tags: [VISA Card]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  format: decimal
      responses:
        '200':
          description: VISA card balance updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  balance:
                    type: number
                    format: decimal
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
        '403':
          description: Profile verification required for this operation
        '404':
          description: VISA card not found
  /visa-card/{id}/check-spending-limits:
    post:
      summary: Check VISA card spending limits
      tags: [VISA Card]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  format: decimal
      responses:
        '200':
          description: Spending limits checked
          content:
            application/json:
              schema:
                type: object
                properties:
                  canSpend:
                    type: boolean
                  dailyRemaining:
                    type: number
                    format: decimal
                  monthlyRemaining:
                    type: number
                    format: decimal
                  requestedAmount:
                    type: number
                    format: decimal
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
        '403':
          description: Profile verification required for this operation
        '404':
          description: VISA card not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Our API JWT token obtained from /auth/web3auth/validate
  schemas:
...
    VISACardResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        cardNumber:
          type: string
        cardType:
          type: string
          enum: [physical, virtual]
        status:
          type: string
          enum: [pending, active, suspended, cancelled]
        balance:
          type: number
          format: decimal
        dailyLimit:
          type: number
          format: decimal
        monthlyLimit:
          type: number
          format: decimal
        createdAt:
          type: string
          format: date-time
        activatedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    CreateVISACardRequest:
      type: object
      required: [cardType, dailyLimit, monthlyLimit]
      properties:
        cardType:
          type: string
          enum: [physical, virtual]
        dailyLimit:
          type: number
          format: decimal
          minimum: 100
        monthlyLimit:
          type: number
          format: decimal
          minimum: 1000
    UpdateVISACardRequest:
      type: object
      properties:
        cardType:
          type: string
          enum: [physical, virtual]
        dailyLimit:
          type: number
          format: decimal
          minimum: 100
        monthlyLimit:
          type: number
          format: decimal
          minimum: 1000
...